tests: Move config loggin to the root index function and add more tests for runAll
