+++ src/index.js
@@ -1,11 +1,12 @@
 /* eslint no-console: 0 */
-/* eslint no-process-exit: 0 */
 /* eslint import/no-dynamic-require: 0 */
 
 'use strict'
 
 const appRoot = require('app-root-path')
 const cosmiconfig = require('cosmiconfig')
+const stringifyObject = require('stringify-object')
+const getConfig = require('./getConfig')
 const printErrors = require('./printErrors')
 const runAll = require('./runAll')
 
@@ -19,29 +20,43 @@ if (process.stdout.isTTY) {
   process.env.FORCE_COLOR = true
 }
 
-cosmiconfig('lint-staged', {
-  rc: '.lintstagedrc',
-  rcExtensions: true
-})
-  .then(result => {
-    // result.config is the parsed configuration object
-    // result.filepath is the path to the config file that was found
-    runAll(packageJson, result.config)
-      .then(() => {
-        // No errors, exiting with 0
-        process.exit(0)
-      })
-      .catch(error => {
-        // Errors detected, printing and exiting with non-zero
-        printErrors(error)
-        process.exit(1)
-      })
+/**
+ * Root lint-staged function that is called from .bin
+ */
+module.exports = function lintStaged() {
+  cosmiconfig('lint-staged', {
+    rc: '.lintstagedrc',
+    rcExtensions: true
   })
-  .catch(parsingError => {
-    console.error(`Could not parse lint-staged config.
+    .then(result => {
+      // result.config is the parsed configuration object
+      // result.filepath is the path to the config file that was found
+      const config = getConfig(result.config)
+
+      if (config.verbose) {
+        console.log(`
+Running lint-staged with the following config:
+${stringifyObject(config)}
+`)
+      }
+
+      runAll(packageJson, result.config)
+        .then(() => {
+          // No errors, exiting with 0
+          process.exitCode = 0
+        })
+        .catch(error => {
+          // Errors detected, printing and exiting with non-zero
+          printErrors(error)
+          process.exitCode = 1
+        })
+    })
+    .catch(parsingError => {
+      console.error(`Could not parse lint-staged config.
 Make sure you have created it. See https://github.com/okonet/lint-staged#readme.
 
 ${parsingError}
 `)
-    process.exit(1)
-  })
+      process.exitCode = 1
+    })
+}
